backend/requirements.txt
asgiref==3.11.1
dj-database-url==3.1.2
Django==6.0.2
Pillow==11.0.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dotenv==1.2.1
sqlparse==0.5.5
tzdata==2025.3
deep-translator==1.11.4